+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * feat: add usage_statistics app to store users' and patients' activity statistics (!1100)
 * feat: add management command for populating statistics to DailyUserAppActivity and DailyUserPatientActivity models (!1133)
 * feat: finalize management command for sending five types of databank data (!1180)
+* feat: add usage statistics export page for downloading statistical data (!1332)
 
 ### Documentation